changed some variable data types, still SWIG troubleshooting

+++ include/fhmanet/fh_channel_message_strobe.h
@@ -61,21 +61,21 @@ namespace gr {
 
       typedef boost::shared_ptr<fh_channel_message_strobe> sptr;
       
-      static sptr make(float period_ms,
+      static sptr make(unsigned int period_ms,
 						double center_freq, 
-						float channel_width, 
+						unsigned int channel_width, 
 						unsigned int num_channels, 
 						unsigned int sequence_length, 
 						unsigned int tx_security_key);
 
-      virtual void set_period(float period_ms) = 0;
-      virtual float period() const = 0;
+      virtual void set_period(unsigned int period_ms) = 0;
+      virtual unsigned int period() const = 0;
       
       virtual void set_center_freq(double center_freq) = 0;
       virtual double center_freq() const = 0;
 
-      virtual void set_channel_width(float channel_width) = 0;
-      virtual float channel_width() const = 0;
+      virtual void set_channel_width(unsigned int channel_width) = 0;
+      virtual unsigned int channel_width() const = 0;
 
       virtual void set_num_channels(unsigned int num_channels) = 0;
       virtual unsigned int num_channels() const = 0;
